[AutoSparkUT] Combine three existing UT migration PRs into one to facilitate CI pipeline (#13811)
### Description

This is a combination of 3 exitsing PRs:

- https://github.com/NVIDIA/spark-rapids/pull/13766
- https://github.com/NVIDIA/spark-rapids/pull/13770
- https://github.com/NVIDIA/spark-rapids/pull/13776

There might be merge conflict if we merge them one by one as they are
modifying the exact same files and adjacent lines, so putting them
together in just one PR is a more friendly way to get tests in.

### Checklists


- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Allen Xu <allxu@nvidia.com>
Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>