Fixes an issue were we were not properly checking log params and nullable output (#13818)
This fixes #13741 
This fixes #13740 
This fixes #13738 
This fixes #13737 

This partially fixes https://github.com/NVIDIA/spark-rapids/issues/89
(at least I think it does)

The comes down to a few issues that were fixed.

1) `log` and `conv` were not properly setting `nullable`. So if the
output ended up with a NULL in it and we tried to bring it back to the
CPU we might have gotten data corruption.
2) `log` had not implemented the all scalar version so it ended up not
getting any validity checks
3) `conv` had implemented an all scalar version, but it threw an
exception. It should almost never happen, so I made a version that does
the right thing.
4) log has only done half of the input checks. This resulted in some
NaN's being returned when Spark wanted them to be nulls.
5) The log check issues were then hidden by the test for them being
marked as xfail. It was 5 years ago when I filed it so I don't know if
things changed or what, but it all appears to pass now.

### Checklists

- [X] This PR has added documentation for new or modified features or
behaviors.
- [X] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

I didn't do any performance testing because correctness trumps
performance. The change should be small because we are doing a few extra
checks, but they are not that much.

---------

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>