Support GpuMapFromEntries to prevent fallback 25.10 [databricks] (#13831)
This cherry picks https://github.com/NVIDIA/spark-rapids/pull/13806 to
25.10

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Alessandro Bellina <abellina@nvidia.com>