Support GpuMapFromEntries to prevent fallback [databricks] (#13806)
Fixes https://github.com/NVIDIA/spark-rapids/issues/9079

### Description

This adds GpuMapFromEntries to prevent fallbacks for what is a very
simple operation. I used mostly AI to write this, and test it, so there
could be some issues i have not found yet.

### Checklists

- [x] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Alessandro Bellina <abellina@nvidia.com>