Add different join strategies, join logging, and heuristic confs (#13717)
This adds in configs that lets a user select a join strategy, Sort merge
Join, Hash Join, or Hash Join with post processing. This is here as a
first step towards a heuristic that can let us automatically select the
strategy to use.

In general this does not change the default performance as it is using
the same algorithm by default, but it lets us address performance issues
when a join could potentially explode, in which case if configured
properly it could be hundreds of times faster than what we have today.

### Description

This also adds in a config to help with debugging if a join is slow, and
a config to tune a heuristic in spilling a join gather map into smaller
parts.

This depends on https://github.com/NVIDIA/spark-rapids-jni/pull/3917

### Checklists

- [X] This PR has added documentation for new or modified features or
behaviors.
- [X] This PR has added new tests or modified existing tests to cover
new code paths.
- [X] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>