Upgrade ucx to 1.19.1-rc2 (#13719)
Fixes https://github.com/NVIDIA/spark-rapids/issues/13266

### Description

Update all ucx version references in the repo to point to 1.19.1-rc2,
and update the corresponding UCX_CUDA_VER to Cuda13, because the new
release contains a rocky cuda13 version.

Note: We are also changing the rocky dockerfiles to use the "rocky"
archives now instead of centos, since that's the one that has cuda13.

### Checklists

Manually tested building the dockerfiles and running `ucx_perftest -m
cuda`

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Zach Puller <zpuller@nvidia.com>