Supports year/month/day/hour partition transforms for Iceberg (#13726)
closes #13384
closes #13385
closes #13386
closes #13387

### Description

Supports year/month/day/hour partition transfroms

### depends on
* https://github.com/NVIDIA/spark-rapids-jni/pull/3903

### Checklists

- [x] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [x] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.
Need to do a end to end testing, perf testing is not in this PR.

---------

Signed-off-by: Chong Gao <res_life@163.com>
Co-authored-by: Chong Gao <res_life@163.com>