Support left-outer joins with no columns in either build or stream side (#13817)
close https://github.com/NVIDIA/spark-rapids/issues/13731

As reported in the linked issue, the `GpuBroadcastNestedLoopJoinExec`
will complain errors when meeting the degenerate left-outer joins as
below. (aka there are no columns in either build or stream side)
```
Caused by: java.lang.ArrayIndexOutOfBoundsException: Index 0 out of bounds for length 0
        at ai.rapids.cudf.Table.<init>(Table.java:58)
        at com.nvidia.spark.rapids.GpuColumnVector.from(GpuColumnVector.java:591)
        at org.apache.spark.sql.rapids.execution.ConditionalNestedLoopJoinIterator.$anonfun$computeNumJoinRows$2(GpuBroadcastNestedLoopJoinExecBase.scala:227)
        at scala.runtime.java8.JFunction0$mcJ$sp.apply(JFunction0$mcJ$sp.scala:17)
```

This is because the cuDF `Table` requires at least one column as the
input, then the current code can NOT support the degenerate cases.
This PR adds in the support by performing the similar join logic by
leveraging the existing Java APIs as the Spark nested-loop left-outer
join for the degenerate cases.

---------

Signed-off-by: Firestarman <firestarmanllc@gmail.com>
Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>