Use pyspark-client for the Spark Connect smoke test (#13778)
Fixes #13718 

### Description

Previous assumption that pyspark[connect] is the smallest pip package
for Connect is wrong.
It has a full SPARK_HOME including the jars. 
Smoke test should be a pure Python app. 

This PR proposes to use pyspark-client as in
[spark-rapids-examples](https://github.com/NVIDIA/spark-rapids-examples/blob/517158bb274400c928aac2772a4e27309095a24a/examples/spark-connect-gpu/client/requirements.txt#L21)


### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: timl <timl@nvidia.com>
Co-authored-by: Tim Liu <timl@nvidia.com>