Replace apache.commons.StringUtils.abbreviate calls
This commit replaces calls to
`apache.commons.lang3.StringUtils.abbreviate` with a custom
implementation from the plugin.

Apache Spark 4.x is reducing its dependency on Apache commons-lang3,
as is evident from the changes made in Apache Spark, in the
following JIRAs:
1. https://issues.apache.org/jira/browse/SPARK-53004
2. https://issues.apache.org/jira/browse/SPARK-52987

Another alternative to this change could have been to use the Apache
Spark implementation of `SparkStringUtils`.  But that implementation is
not available till Apache Spark v4.1, and would require a shim to use.
This change is simpler to make.

Signed-off-by: MithunR <mithunr@nvidia.com>