Use shim to identify whether DataWriting is supported for LoRe (#13857)
Follow-up https://github.com/NVIDIA/spark-rapids/pull/13819

### Description

We don't support the `GpuDataWritingCommandExec` for LoRe since
`WriteFiles` was introduced in Spark 3.4.0. Currently, we check if it is
supported using a hardcoded list, which makes it easy to forget to
update it when a new shim layer is added.

This PR moves the check to the shim layer, so we no longer need to
remember to sync the list.

I thought this PR would fix
https://github.com/NVIDIA/spark-rapids/issues/13855, but it turns out
that it's a duplicate of
https://github.com/NVIDIA/spark-rapids/issues/13798 that has already
been fixed by @jihoonson in
https://github.com/NVIDIA/spark-rapids/pull/13819. I didn't upmerge my
code. Thanks!

### Checklists

- [x] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

Signed-off-by: Haoyang Li <haoyangl@nvidia.com>