Use pyspark-client only for Spark 4.x [skip ci] (#13867)
Fixes #13860 

### Description

pyspark-client is available only for Spark 4.x, use pyspark[connect] for
3.5.6+ tests in the smoke test

### Testing

```bash
$ SPARK_HOME=~/dist/spark-4.0.0-bin-hadoop3  SCALA_VERSION=2.13 \
  SPARK_CONNECT_SMOKE_TEST=1 ./integration_tests/run_pyspark_from_build.sh

$ SPARK_HOME=~/dist/spark-3.5.6-bin-hadoop3  \
  SPARK_CONNECT_SMOKE_TEST=1 ./integration_tests/run_pyspark_from_build.sh 
```

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: timl <timl@nvidia.com>
Signed-off-by: Gera Shegalov <gshegalov@nvidia.com>
Co-authored-by: Tim Liu <timl@nvidia.com>