Fix test_delta_merge_sql_liquid_clustering for Delta 4 [databricks] (#13849)
Fixes #13848.

### Description

The test in question is failing because of the `HashAggregateExec` in
one of metadata processing queries. Delta 4 seems to generate a
different query plan for this metadata query from Delta 3.3.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Jihoon Son <ghoonson@gmail.com>