Ignore file metadata in DV-enabled Delta Scan to enable GPU by default (#13843)
Contributes to #7458 

### Description

This PR proposes to ignore file metadata in Delta Scan with DV because
it is not used in the plan and to prevent fallbacks in this case

This enables GPU execution of DV Scans by default using the same
algorithm that previously required setting useMetadataRowIndex = false


### Checklists


- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Gera Shegalov <gshegalov@nvidia.com>
Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>