Persist the buffer converters of TypedImperativeAggregate into LogicalPlan [databricks] (#13875)
Fixes #13820

This PR attempts to fix the subtle issue (#13820) of failing to convert
aggregate buffers of `TypedImperativeAggregate` from CPU/GPU to the
counterpart when AQE is enabled. Specifically, this bug is due to agg
buffer converters, which are inserted into `SparkPlan` as `TreeNodeTag`,
will be lost if the re-optimize of AQE takes effect (by then, the plan
will be totally rebuilt on the top of the new logical plan).

---------

Signed-off-by: sperlingxx <lovedreamf@gmail.com>