Remove low priority iceberg it test cases to save test time [skip ci] (#13922)
Related to #13881  #13885 

### Description

We removed a lot of low priority tests cases to reduce iceberg
integration tests time. On my local desktop, its time reduced from 5.5h
to 1.33h.

### Checklists

- [x] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [x] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: liurenjie1024 <liurenjie2008@gmail.com>