Add in API definitions for Rapids UDAF (#13870)
Contributes to https://github.com/NVIDIA/spark-rapids/issues/13412

Rapids UDAF is designed to support executing an UDAF (User Defined
Aggregate Function) in the columnar way to get accelerated by GPU.

Complete support of RapidsUDAF covers too many things and a single PR
(https://github.com/NVIDIA/spark-rapids/pull/13450) is too large to
review. So instead it's better to be added in piece by piece, and this
PR is the first one who only introduces the relevant inerfaces.

- `RapidsUDAF`- the top interface, it defines 5 methods as below, trying
to follow the [CPU definitions
](https://github.com/apache/spark/blob/v3.5.0/sql/core/src/main/scala/org/apache/spark/sql/expressions/udaf.scala#L35)(`UserDefinedAggregateFunction`)
as much as possible to minimize users' learning effort.

   | RapidsUDAF |  UserDefinedAggregateFunction | 
   | :---- | :---- |
   | getDefaultValue  | initialize  |
   | updateAggregation | update |
   | mergeAggregation | merge |
   | getResult | evaluate |
   | bufferTypes | bufferSchema |

`updateAggregation` and `mergeAggregation` return a
`RapidsUDAFGroupByAggregation` who contains the APIs to perform the
aggregation.
- `RapidsUDAFGroupByAggregation` - base interface for GPU-accelerated
UDAF aggregation implementations. It provides the contract for different
aggregation strategies. it also supports an optional pair of `preStep`
and `postStep` to run some transformations before and after a
"reduce/aggregate" operation, similar as "preMerge" and "postMerge" for
the merge-stage aggregate in `GpuAggregateFunction`.
- `RapidsSimpleGroupByAggregation` - the child class of
`RapidsUDAFGroupByAggregation`, providing a standard cuDF-based
aggregation step that uses built-in cuDF aggregation operations.

Putting the groupby 'aggregate' API in the child class is because more
types of `aggregate` may be introduced in the future via child classes.
e.g. an `aggregate` as below can access the grouped data and keys to let
users do more customization.
```
  /**
   * Performs custom aggregation on data that has been grouped by keys.
   * The data is grouped, with offsets indicating group boundaries.
   * @param keyOffsets A ColumnVector containing the start offset for each group.
   *                   The end offset for group i is `keyOffsets[i+1]` (or total
   *                   rows for the last group).
   * @param groupedData An array of ColumnVectors containing the actual data
   *                    columns, sorted and organized by the grouping keys.
   * @return An array of ColumnVectors with one row per group, containing the
   * aggregated results.
   */
  ColumnVector[] aggregateGrouped(ColumnVector keyOffsets, ColumnVector[] groupedData);
```

---------

Signed-off-by: Firestarman <firestarmanllc@gmail.com>
Co-authored-by: greptile-apps[bot] <165735046+greptile-apps[bot]@users.noreply.github.com>