Fix the ordering of join gather maps, and make it harder to break (#13929)
This fixes #13854 

I thought I had tested this completely, but clearly I had not. I am
putting this up here, but I plan to run a lot more tests.

The issue was that I had made a mistake on one API where the gather maps
returned were in the wrong order and that resulted in us trying to
gather data that was out of bounds. What is worse it was for an inner
join so it resulted in us walking off the end of the data and reading
bogus data, which might be the reason for the hang.

### Description

This code is larger than it needs to be. I wanted to put something in
that would ensure that we did this right and that we would not run into
issues again in the future for this same issue. That said there are a
lot of cases where the build side of the join is implied, especially by
the AST condition. It would be nice to fix this too, but it is beyond
the scope of this issue.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>