Fix a special case in limit where it could return an empty batch with the wrong number of columns (#13938)
This Fixes #13914

### Description

The code was wrong, but not so wrong that unless we broadcast the empty
batch it would probably just be filtered out and ignored.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [X] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>