Make the GPU UDF has different name than the CPU one (#13970)
This PR adds the GPU specific prefix to the UDF name to make it differ
from the CPU one.

Some customers want to know if the
[GpuRowBasedUDF](https://github.com/NVIDIA/spark-rapids/blob/main/sql-plugin/src/main/scala/org/apache/spark/sql/rapids/GpuScalaUDF.scala#L94)
is applied from the physical plan in the eventlog other than the explain
log in the driver. But the GPU UDF has the same name as the CPU, so it
is not easy to achieve this for their tools. This PR will add this type
of support.

Signed-off-by: Firestarman <firestarmanllc@gmail.com>