Fix a join where only a single column is used as a condition (#13964)
This Fixes #13907

### Description

The test case exposed two bugs

1. An issue where the AST spilling code was not doing the right thing
when a top level expression could not be converted to AST
2. A right outer degenerate join was not being processed correctly.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [X] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>