Fix join bug on csv datasources (#13903)
fixes #13873

### bug analysis

This is a corner case described in the issue #13873.
If reading other type datasources instead of CSV files, the error does
not occur.
If using `.collect()` instead of `.show()`, the error does not occur.
If the stream data is not empty, the error does not occur.

CSV file with only header and no data rows, empty table(num_rows = 0)
will be returned, then after project, the column number of empty table
becomes 0, finally empty table(num_rows=0, num_cols=0) occurs in the
iterator.

### bug fix
If num_rows of table is 0 when reading csv file, return None instead of
empty table, the None indicates the iterator of CSV data is empty.

Signed-off-by: Chong Gao <res_life@163.com>

---------

Signed-off-by: Chong Gao <res_life@163.com>
Co-authored-by: Chong Gao <res_life@163.com>