Add in support for getting SQL metrics from expressions (#13981)
This is part 1 of splitting
https://github.com/NVIDIA/spark-rapids/pull/13368 into smaller parts

### Description

This adds in framework changes to allow an expression to update SQL
metrics. The metrics are passed in when the expressions are bound and
then can be updated as they run. There are no expressions added, yet
that would let us truly test it beyond seeing that it does not break
existing functionality.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

This is an internal only feature so no docs or tests are really needed,
beyond verifying that we didn't break anything.

Most of the changes are very small, except for the ones in
GpuBoundAttribute and GpuMetrics

---------

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>