Add layer of indirection when converting expressions to the GPU (#13987)
This is part 2 of splitting up
https://github.com/NVIDIA/spark-rapids/pull/13368

### Description

This adds in a layer of indirection when converting an expression to be
on the GPU. It is not used in the PR, but in the final PR it will allow
us to wrap CPU expressions in a GPU expression compatibility layer
(bridge).

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

No real need for new tests. If they still all pass and everything
compiles, then the refactoring is good.

---------

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>