Building scala213 plugin for spark350+ (#13992)
Part of #13751 .

### Description

1. Update scala2.13 CI build + test to cover spark3.5.0+ only.
2. Remove also CDH profiles

Note: the actual cleanup of code and doc should be covered in other
changes.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [X] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

Signed-off-by: Peixin Li <pxLi@nyu.edu>