Add pmattione to blossom: Attempt 2  [skip ci] (#13997)
Adds myself to blossom-ci.yml to trigger builds.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Paul Mattione <pmattione@nvidia.com>