Drop _tmp_metadata_row_index column from the output of Delta Scan on GPU (#13991)
### Description

When returning output from `GpuFileSourceScanExec`, we leave the
temporary row_index column that is generated as part of the DV scan. The
presence of this column breaks our assumption that partition columns
follow immediately after the output column. This PR removes the
`_tmp_metadata_row_index` column from the output.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Raza Jafri <raza.jafri@gmail.com>