Add parquet mixed-encodings test (#13982)
Fixes #[13739](https://github.com/NVIDIA/spark-rapids/issues/13739).

### Description

This adds an integration test to test writing and reading parquet
columns with mixed encodings. We write a string column with a few pages
of repeated short strings, and a with a few pages of extremely long
unique strings. We test this with both the cpu and gpu parquet writers.

Both writers fall back to plain encoding for the pages with long unique
strings. The gpu writer puts the data with different encodings in
different row groups, whereas the cpu writer keeps them in the same row
group. This test then makes sure that both the cpu & gpu reads of both
the cpu- & gpu-generated data are identical.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Paul Mattione <pmattione@nvidia.com>