Skip test_delta_filter_out_metadata_col on [databricks]  (#14002)
### Description

A recently introduced test in #13991 is irrelevant for DBR as we
currently don't support Deletion Vectors on DBR. This PR skips the test
on DBR and for versions of Spark that don't support Deletion Vectors

### Checklists


- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Raza Jafri <raza.jafri@gmail.com>