Delta table scan should be optimized when deletion vectors don't exist [databricks] (#13980)
Fixes https://github.com/NVIDIA/spark-rapids/issues/13950 and
https://github.com/NVIDIA/spark-rapids/issues/14004.

### Description

Currently, when reading on GPU , the plugin cannot support certain
optimizations for Delta tables such as file split or predicate pushdown
if the deletion vectors are enabled for them. This change was made in
https://github.com/NVIDIA/spark-rapids/pull/13843. However, we
accidentally disabled the optimizations even for Delta tables without
deletion vectors. This PR fixes this regression by disabling the
optimization for Delta tables if they have deletion vectors.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Jihoon Son <ghoonson@gmail.com>