Fixes various issues in 25.12 release branch [databricks] (#14013)
Fixes #14009 

### Description

There are three issues:

1. `test_delta_scan_split_with_DV_*` tests should run with Spark 3.5.3+
only.
2. `test_delta_scan_split_with_DV_enabled_after_DVs_materialized` test
should be skipped with Spark 4.0.1 due to
https://github.com/delta-io/delta/issues/5690.
3. `test_delta_filter_out_metadata_col` flakes as the query plan
sometimes does not involve reading deletion vectors. This seems to
happen when there are only a few rows deleted.

I manually tested `delta_lake_test.py` with Spark 3.2.4, 3.3.4, 3.4.4,
3.5.7, 4.0.0, and 4.0.1.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Jihoon Son <ghoonson@gmail.com>