SHA-2 hash support.
This commit adds support for the [SHA-2
family](https://spark.apache.org/docs/latest/api/sql/index.html#sha2) of hash functions.

As with Apache Spark, the following data types are supported:
1. `BINARY`
2. `STRING`

Also, as with Apache Spark, the following bit-lengths are supported:
1. 224
2. 256
3. 384
4. 512
If the bitlength is specified as 0, a default of 256 is used.
If the bitlength differs from the list above, the returned row is null.

This implementation differs from Apache Spark in that it requires the
bitlength to be an INTEGER literal.  Apache Spark supports expressions
as bitlength, e.g.
```sql
SELECT SHA2(my_str, LENGTH(my_str)) FROM ...
```
This might be an exceedingly rare usecase, and can be addressed at a
later date if deemed a priority.

Signed-off-by: MithunR <mithunr@nvidia.com>