support withRetry with split for shuffle exchange exec base (#13975)
Fixes https://github.com/NVIDIA/spark-rapids/issues/13951 (along with
https://github.com/NVIDIA/spark-rapids/pull/14010)

### Description

We convert the internal state of `prepareBatchShuffleDependency` in
`rddWithPartitionIds` into a spliterator to support split retires. This
happens a level above where we call into the partitioner which does the
gpu kudo serialization.

Performance testing so far shows no significant change for low memory
scenarios in NDS. I'll continue with perf testing while the PR is under
review.

### Checklists


- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [x] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Zach Puller <zpuller@nvidia.com>