Update documentation to reflect the shims updates [skip ci] (#14058)
Fixes #13751

### Description

This change updates the documentation to reflect the behavior introduced
in #13751.

NOTE: We are not actually removing the deprecated shims from the
codebase yet. That cleanup should be handled by the project maintainers
in the future.

### Checklists
- [X] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

Signed-off-by: Peixin Li <pxLi@nyu.edu>