Add Iceberg view test (#14042)
closes #13935.

### Description

Add Iceberg view tests.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
No need to update document. View creation/drop runs on CPU. Query on
view can run on GPU if the sql in the view is spoorted by GPU.

- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)

- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.
No need to do performance test.

---------

Signed-off-by: Chong Gao <res_life@163.com>
Co-authored-by: Chong Gao <res_life@163.com>