[BUG] Fix initialization order NPE for RapidsShuffleManager in UCX mode for Spark 4+ [databricks] (#14033)
### Description

This PR changes the heartbeat manager such that it doesn't require the
shuffle manager to be initialized, and defers the registration to the
heartbeat mechanism, or when the shuffle manager is initialized if
heartbeats are disabled (not the default).

Tested this with spark 4.0.1, and spark 3.5.4 in UCX and Multi-Threaded
mode, with and without heartbeats enabled.

Fixes https://github.com/NVIDIA/spark-rapids/issues/13629
Fixes https://github.com/NVIDIA/spark-rapids/issues/14034
### Checklists

- [x] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Alessandro Bellina <abellina@nvidia.com>